requirements.txt
requests
python-dotenv
pyodbc
bs4